+++ src/index.ts
@@ -1,3 +1,2 @@
-import ChapterMap from "./ChapterMap";
-
-export default ChapterMap;
+export { default, defaultGeographyFilter } from "./ChapterMap";
+export type { ChapterMapProps } from "./ChapterMap";
